fix: Small bug fixes, and code cleaning, commenting
