+++ globals.py
@@ -1,6 +1,13 @@
-
+# All global variables are in this file.
 
 def init():
-    global LOOP, RECORDER
-    LOOP = None
-    RECORDER = None
+    global recorder, loop, searching_for_face, idle_skill, waiting_for_response, first_contact, return_to_idle, \
+        length_of_audio
+    loop = None
+    recorder = None
+    searching_for_face = None
+    idle_skill = None
+    waiting_for_response = None
+    first_contact = None
+    return_to_idle = None
+    length_of_audio = None
