+++ requirements.txt
@@ -10,3 +10,5 @@ cvzone==1.5.6
 mediapipe==0.8.10.1
 PyAudio==0.2.11
 
+
+pillow~=9.2.0
